add mock variables to fixtures folder
